In-progress: decoupling the Game class from everything else.
