+++ src/arclia/kerblaxion/__main__.py
@@ -1,6 +1,8 @@
 
 from .game import Game
+from .hero import prepare
 
 def main():
   game = Game()
+  prepare(game)
   game.run()
